TranslationGizmo: introduced (WIP), introduced closest-point onto ray from ray, introduced DynamicGeometry buffer

+++ src/slic3r-shared/include/Slic3r/App/Scene/NodeBuilder.hpp
@@ -55,7 +55,7 @@ public:
         return *this;
     }
 
-    NodeBuilder& set_screen_space_sized_modifier(float scale);
+    NodeBuilder& set_screen_space_sized_modifier(double scale);
 
     NodeBuilder& set_debug_name(std::string_view sv)
     {
