+++ src/slic3r-shared/src/Slic3r/App/Scene/AabbRaycastNodeComponent.cpp
@@ -9,7 +9,7 @@ bool AabbRaycastNodeComponent::raycast(const Matrix4d& world, const Ray& ray, do
 {
     ASSERT(m_aabb_mesh != nullptr);
 
-    Matrix4d inv_world = world.cast<double>().inverse();
+    Matrix4d inv_world = world.inverse();
 
     Vec3d local_ray_origin = (inv_world * Vec4d{ray.origin.x(), ray.origin.y(), ray.origin.z(), 1}).head<3>().cast<double>();
     Vec3d local_ray_direction = (inv_world.block<3, 3>(0, 0) * ray.direction).cast<double>().normalized();
